feat: add `--version` flag
